backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
supabase==2.3.5
gotrue==1.3.0
httpx==0.24.1
pydantic[email]==2.6.3
pydantic-settings==2.2.1
websockets==12.0